refactor : repository 내부 클래스명 변경
