+++ src/main/java/roomescape/service/ReservationTimeService.java
@@ -4,31 +4,31 @@ import java.util.List;
 import org.springframework.stereotype.Service;
 import roomescape.controller.request.ReservationTimeRequest;
 import roomescape.model.ReservationTime;
-import roomescape.repository.ReservationTimeDAO;
+import roomescape.repository.ReservationTimeRepository;
 
 @Service
 public class ReservationTimeService {
 
-    private final ReservationTimeDAO reservationTimeDAO;
+    private final ReservationTimeRepository reservationTimeRepository;
 
-    public ReservationTimeService(ReservationTimeDAO reservationTimeDAO) {
-        this.reservationTimeDAO = reservationTimeDAO;
+    public ReservationTimeService(ReservationTimeRepository reservationTimeRepository) {
+        this.reservationTimeRepository = reservationTimeRepository;
     }
 
     public List<ReservationTime> findAllReservationTimes() {
-        return reservationTimeDAO.findAllReservationTimes();
+        return reservationTimeRepository.findAllReservationTimes();
     }
 
     public ReservationTime addReservationTime(ReservationTimeRequest request) {
         ReservationTime reservationTime = new ReservationTime(request.startAt());
-        return reservationTimeDAO.addReservationTime(reservationTime);
+        return reservationTimeRepository.addReservationTime(reservationTime);
     }
 
     public ReservationTime findReservationTime(long id) {
-        return reservationTimeDAO.findReservationById(id);
+        return reservationTimeRepository.findReservationById(id);
     }
 
     public void deleteReservationTime(long id) {
-        reservationTimeDAO.deleteReservationTime(id);
+        reservationTimeRepository.deleteReservationTime(id);
     }
 }
