+++ src/test/java/roomescape/repository/ReservationTimeDAOTest.java
@@ -19,7 +19,7 @@ import roomescape.model.ReservationTime;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
 @DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
-class ReservationTimeDAOImplTest {
+class ReservationTimeDAOTest {
 
     @Autowired
     JdbcTemplate jdbcTemplate;
@@ -28,7 +28,7 @@ class ReservationTimeDAOImplTest {
     DataSource dataSource;
 
     @Autowired
-    ReservationTimeDAOImpl reservationTimeDAOImpl;
+    ReservationTimeDAO reservationTimeDAOImpl;
 
     SimpleJdbcInsert insertActor;
